Revert "Bugfix: it is intended that the argument must be null in this API call."
This reverts commit 8e41527860d002e340ccd4981376cadb540e31d7.
Apparently today there has been a glitch in Dropbox APIs or servers, in
which the request performed for usersGetAccountInfo was not recognized
as properly formed unless the body contains some JSON data. Being a
method that takes no argument 'null' was the natural choice. I then
realized that it was not sent at all in the request nonetheless because
I was looking at the old API v1 documentation. Double checking now shows
that it is working correctly (which is a relief).