Bugfix: dynamic formula graph update with reindex.
The id and path need to be all updated when reconnecting predecessors
and successors because when we match relative selectors, we compare the
path.

Moreover, the selector instances must also have an updated list of the
formula nodes using the selectors, otherwise the reverse match will not
yield any match for that specific formula node.

Thirdly, we cache the selector instances in the formula node but we
delete them from the selector storage: this puts the sets out of sync,
so we recycle the cached selector instance.